Move non-async methods to extensions
